Restructure codebase: move all codecs into new package persistence.codec
